requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap4==2024.10
Django==5.1.2
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-environ==0.11.2
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
lxml==5.3.0
pillow==11.0.0
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-docx==1.1.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xlsxwriter
pycryptodome==3.19.0
python-dateutil==2.8.2
django-storages
boto3
reportlab
